feat: benchmark video segmentation
feat: paths to videos point to new archive

fix: add untracked test file

feat: clean-target more explicit

feat: use secure filenames on archive vids

feat: segmentation benchmarks

fix: repair broken indentation in Makefile

fix: repair empty dev-requirements file

pipenv lock -d and pipenv lock --dev is not
adequate. It must be pipenv lock --requirements --dev

fix: infinite job TTL required for long videos

fix: REACT_APP_API_URL no longer version controlled

fix: repair environment variables referencing directories

+++ video_reuse_detector/ffmpeg.py
@@ -39,16 +39,22 @@ def extract_outputs(log_file: str) -> List[Path]:
     return list(filter(lambda path: path.exists(), all_paths))
 
 
-def execute(cmd: str, output_directory: Path) -> List[Path]:
+def execute(cmd: str, output_directory: Path, remove_log=True) -> List[Path]:
     if not output_directory.exists():
-        logger.debug(f'Creating "{output_directory}" and parents if necessary')
+        logger.debug(f'Creating "{output_directory}" (and parents) if necessary')
         output_directory.mkdir(parents=True)
-
-    logger.debug(f'Executing: "{cmd}"')
+    else:
+        logger.debug(f'"{output_directory}" exists')
 
     # TODO: Use tempfile?
     log_file = f'ffreport{random.randint(0, 1000)}.log'
     ffmpeg_env = {'FFREPORT': f'file={log_file}:level=48'}
+
+    if not os.access(str(output_directory), os.X_OK | os.W_OK):
+        logger.error(f'Do not have write access to {output_directory}')
+        raise PermissionError
+
+    logger.debug(f'Executing: "{cmd}", logging to "{log_file}"')
 
     subprocess.call(
         cmd.split(),
@@ -59,7 +65,9 @@ def execute(cmd: str, output_directory: Path) -> List[Path]:
 
     try:
         output_paths = extract_outputs(log_file)
-        os.remove(log_file)
+
+        if remove_log:  # Usually True, but handy for debugging
+            os.remove(log_file)
     except FileNotFoundError:
         msg = (
             f'Could not read/remove log file "{log_file}"'
@@ -67,7 +75,10 @@ def execute(cmd: str, output_directory: Path) -> List[Path]:
         )
         logger.warning(f'${msg}. log_file="{log_file}"')
 
-    logger.info(f'Produced output files: "{format_outputs(output_paths)}"')
+    if output_paths == []:
+        logger.warning(f'Executing \"{cmd}\" did not produce any output!')
+    else:
+        logger.info(f'Produced output files: "{format_outputs(output_paths)}"')
 
     return output_paths
 
@@ -90,7 +101,7 @@ def slice(
     file_path: Path, ss: str, duration: str, output_directory: Path, overwrite=False
 ) -> Path:
     # Incoming strings on the form HH:MM:SS turned into HHMMSS
-    suffix = f"{ss.replace(':', '')}_{duration.replace(':', '')}"
+    suffix = f"start_{ss.replace(':', '')}_duration_{duration.replace(':', '')}"
 
     # Concatenate into the original file name, append a delimiting underscore
     # add the extension of the original file, so for,
@@ -102,7 +113,7 @@ def slice(
     # ss='00:00:30'
     # duration='00:00:05'
     #
-    # we get "ATW-550_000030_000005.mpg"
+    # we get "ATW-550_start_000030_duration_000005.mpg"
     extension = file_path.suffix  # includes the ".", so for instance ".mpg"
     name_of_new_file = f'{file_path.stem}_{suffix}{extension}'
     output_path = output_directory / Path(name_of_new_file)
@@ -122,4 +133,94 @@ def slice(
         f' -y {str(output_path)}'
     )
 
-    return execute(cmd, output_path.parent)[0]
+    return execute(cmd, output_path.parent, remove_log=False)[0]
+
+
+def __method__():
+    import traceback
+
+    return traceback.extract_stack(None, 2)[0][2]
+
+
+def get_output_file_name(input_file: Path, video_filter: str, params={}) -> Path:
+    extension = input_file.suffix  # includes the ".", so for instance ".mp4"
+
+    name_of_new_file = f'{input_file.stem}_{video_filter}'
+
+    if params != {}:
+        # Join all k, v pairs with an underscore between them
+        suffix = '_'.join(f'{k}_{v}' for k, v in params.items())
+        name_of_new_file = f'{name_of_new_file}_{suffix}'
+
+    name_of_new_file += extension
+
+    return Path(name_of_new_file)
+
+
+# blur (gaussian, motion),
+#
+# See: https://medium.com/@allanlei/blur-out-videos-with-ffmpeg-92d3dc62d069
+def blur(
+    input_file: Path,
+    output_directory: Path,
+    luma_radius=2,
+    chroma_radius=10,
+    luma_power=1,
+) -> Path:
+    params = {
+        'luma_radius': luma_radius,
+        'chroma_radius': chroma_radius,
+        'luma_power': luma_power,
+    }
+
+    output_file_name = get_output_file_name(input_file, __method__(), params)
+    output_path = output_directory / output_file_name
+    logger.debug(
+        f'Applying {__method__()} with params={params} on {input_file} producing {output_path}'  # noqa: E501
+    )
+
+    cmd = (
+        f'ffmpeg -i {input_file}'
+        ' -filter_complex '
+        f'[0:v]boxblur=luma_radius={luma_radius}'
+        f':chroma_radius={chroma_radius}'
+        f':luma_power={luma_power}[blurred]'
+        ' -map [blurred]'
+        f' {output_path}'
+    )
+
+    return execute(cmd, output_path.parent, remove_log=False)[0]
+
+
+def get_frame_at_time(input_file: Path, output_directory: Path, timestamp: str) -> Path:
+    output_file_name = f"{input_file.stem}_{timestamp.replace(':', '')}.png"
+    output_path = output_directory / output_file_name
+    logger.debug(f'Applying {__method__()} on {input_file} producing {output_path}')
+
+    cmd = f'ffmpeg -ss {timestamp} -i {input_file} -vframes 1 {output_path}'
+
+    return execute(cmd, output_path.parent, remove_log=False)[0]
+
+
+def apply_frei0r_filter(
+    input_file: Path, output_directory: Path, video_filter: str, overwrite=False
+) -> Path:
+    assert input_file.exists()
+
+    output_file_name = get_output_file_name(input_file, video_filter)
+    output_path = output_directory / output_file_name
+
+    if not overwrite and output_path.exists():
+        logger.debug(f'{output_path} exists already, returning without calling ffmpeg')
+        return output_path
+
+    logger.debug(f"Adding {video_filter} to {input_file} producing {output_path}")
+
+    return execute(
+        f'ffmpeg -i {input_file} -vf frei0r={video_filter} {output_path}',
+        output_path.parent,
+    )[0]
+
+
+def softglow(input_file: Path, output_directory: Path, overwrite=False) -> Path:
+    return apply_frei0r_filter(input_file, output_directory, __method__(), overwrite)
